requirements.txt

### Contenido del `requirements.txt`

```plaintext
speechrecognition
pyttsx3
pyautogui
wikipedia-api
openai
requests
python-dotenv
